requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
asyncpg==0.29.0
psycopg2-binary>=2.9.5
bcrypt>=4.2.0
jinja2>=3.1.2
python-multipart>=0.0.20
anthropic>=0.75.0
itsdangerous>=2.2.0
authlib>=1.6.6
alembic>=1.17.2
jsonschema>=4.25.1
jsonschema-path>=0.3.4
jsonpath-ng>=1.7.0
sse-starlette>=2.0.0
boto3>=1.35.0
mypy>=1.19.1
pytest-cov>=7.0.0
radon>=6.0.1